test: extract sampling-equivalence helpers into a shared header
test_logical_noise_prefix.cc's random noisy-circuit generator and
cross-binomial sampling-equivalence checker are exactly what a
schedule-dependence test for noise-transparent reordering needs too.
Move both into tests/sampling_equivalence_helpers.h and have
test_logical_noise_prefix.cc pull them in via clifft::test, so the
next commit's new test file can reuse them instead of duplicating
several hundred lines. No behavioral change: the extracted functions
keep their bodies verbatim.
